+++ network-api/networkapi/wagtailpages/pagemodels/buyersguide/forms.py
@@ -1,0 +1,35 @@
+from wagtail.admin.forms import WagtailAdminModelForm, WagtailAdminPageForm
+
+
+class BuyersGuideProductCategoryForm(WagtailAdminModelForm):
+    def clean_name(self):
+        instance = self.instance
+        set = instance.__class__.objects
+        name = self.cleaned_data["name"]
+        duplicate = set.filter(name__iexact=name).exclude(pk=instance.pk)
+
+        if hasattr(instance, "locale"):
+            duplicate = duplicate.filter(locale=instance.locale)
+
+        if duplicate.exists():
+            self.add_error("name", "A category with this name already exists.")
+
+        return name
+
+    def clean_parent(self):
+        parent = self.cleaned_data["parent"]
+        if parent:
+            if parent.parent:
+                self.add_error("parent", "Categories can only be two levels deep.")
+            if self.instance == parent:
+                self.add_error("parent", "A category cannot be a parent of itself.")
+        return parent
+
+
+class BuyersGuideArticlePageForm(WagtailAdminPageForm):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        # Overriding admin form in order to make promote tab's
+        # "search image" and "search description" fields required.
+        self.fields["search_description"].required = True
+        self.fields["search_image"].required = True
